# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/devices": {
            "get": {
                "description": "根据查询条件获取设备列表，并返回分页数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取设备列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤条件，格式为 filter[key]=value, 多个过滤条件用 \u0026 分隔",
                        "name": "filter",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vo.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "根据提供的设备信息进行设备的保存或更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "保存或更新设备",
                "parameters": [
                    {
                        "description": "设备信息",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功!",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据设备ID删除指定的设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功!",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            }
        },
        "/admin/modules": {
            "get": {
                "description": "根据查询条件获取模块列表，并返回分页数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取模块列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤条件，格式为 filter[key]=value, 多个过滤条件用 \u0026 分隔",
                        "name": "filter",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vo.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "根据提供的模块信息进行模块的保存或更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "保存或更新模块",
                "parameters": [
                    {
                        "description": "模块信息",
                        "name": "module",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ModuleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功!",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据模块ID删除指定的模块",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除模块",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "模块ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功!",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            }
        },
        "/admin/products": {
            "get": {
                "description": "根据查询条件获取产品列表，并返回分页数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取产品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤条件，格式为 filter[key]=value, 多个过滤条件用 \u0026 分隔",
                        "name": "filter",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vo.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "根据提供的产品信息进行产品的保存或更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "保存或更新产品",
                "parameters": [
                    {
                        "description": "产品信息",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProductDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功!",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据产品ID删除指定的产品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除产品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "产品ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功!",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            }
        },
        "/admin/properties": {
            "get": {
                "description": "根据查询条件获取属性列表，并返回分页数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取属性列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤条件，格式为 filter[key]=value, 多个过滤条件用 \u0026 分隔",
                        "name": "filter",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vo.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "根据提供的属性信息进行属性的保存或更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "保存或更新属性",
                "parameters": [
                    {
                        "description": "属性信息",
                        "name": "property",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PropertyDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功!",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据属性ID删除指定的属性",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除属性",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "属性ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功!",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "用户通过用户名和密码登录，并返回 token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录信息",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vo.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "用户注册，提供用户名和密码进行注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册信息",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRegisterDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/vo.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.DeviceDTO": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "device_secret": {
                    "type": "string"
                },
                "module_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "dto.ModuleDTO": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "module_name": {
                    "type": "string"
                }
            }
        },
        "dto.ProductDTO": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "product_key": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "product_secret": {
                    "type": "string"
                }
            }
        },
        "dto.PropertyDTO": {
            "type": "object",
            "properties": {
                "data_type": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "property_name": {
                    "type": "string"
                }
            }
        },
        "dto.UserLoginDTO": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UserRegisterDTO": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "vo.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码，200 为成功，其他为失败",
                    "type": "integer"
                },
                "data": {
                    "description": "可选字段，为空时不包含在 JSON 中"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        }
    }
}